Move model tests to Models namespace
